Move Tx-specific interpretation out of common GcmHeader
