+++ src/node/channels.h
@@ -13,6 +13,7 @@
 #include "ds/hex.h"
 #include "ds/serialized.h"
 #include "ds/state_machine.h"
+#include "ds/thread_messaging.h"
 #include "enclave/enclave_time.h"
 #include "entities.h"
 #include "node_types.h"
@@ -59,7 +60,7 @@ namespace ccf
 
   static inline RecvNonce get_nonce(const GcmHdr& header)
   {
-    return RecvNonce(header.get_iv_int());
+    return *reinterpret_cast<const RecvNonce*>(header.iv);
   }
 
   enum ChannelStatus
@@ -193,7 +194,7 @@ namespace ccf
     {
       status.expect(ESTABLISHED);
 
-      RecvNonce recv_nonce(header.get_iv_int());
+      auto recv_nonce = *reinterpret_cast<const RecvNonce*>(header.iv);
       auto tid = recv_nonce.tid;
       assert(tid < threading::ThreadMessaging::max_num_threads);
 
@@ -847,7 +848,8 @@ namespace ccf
         (size_t)nonce.nonce);
 
       GcmHdr gcm_hdr;
-      gcm_hdr.set_iv_seq(nonce.get_val());
+      const auto nonce_n = nonce.get_val();
+      gcm_hdr.set_iv((const uint8_t*)&nonce_n, sizeof(nonce_n));
 
       std::vector<uint8_t> cipher(plain.n);
       assert(send_key);
